fix: unit-tests signal_validate RT constants, cppcheck warnings, AML namespace walk bounds, e2e debug spam
- host_libc test_signal_validate: align SIGRTMAX/SIG_MAX with kernel ABI
  (signals 1-63 valid; kernel pending-signal mask is uint64_t, so signal
  64 cannot exist). Fixes the 3 failing RT-signal assertions.
- acpi_aml: bound all 7 namespace child walks (hop limit + index range)
  so a corrupt/cyclic next_sibling chain can't hang boot in an infinite
  memcmp loop; fix aml_add_node self-parenting off-by-one (count++ ran
  before the parent<count check, allowing parent==idx).
- fat32/bufcache: remove per-sector read debug spam (9152+ lines/boot)
  that flooded serial and starved e2e timing; keep SLOW/error diagnostics.
- e2e.sh: wait for shell module load before running command tests
  (telnetd banner appears before shell.ko registers shell_process_line).
- cppcheck (2.19 clean): dma.h DMA_BIT_MASK inline fn, btrfs/ext4 buffer
  lifetime hoists, sound_oss/virtio_gpu/ext2 uninitvar, hfs/tarfs OOB
  memcpy/memcmp, dlfcn strnlen, rsa unsigned borrow shift, scheduler dst
  init, kernel.c/kunit initcall uintptr_t, kunit FILL_CASES sizeof-bound
  loops, bpf_verifier dead enum collision, rseq NULL guard, sysrq
  suppress, test.c intentional-UB suppressions.

+++ src/fs/initramfs.c
@@ -77,7 +77,7 @@ extern uint8_t _binary_build_initramfs_cpio_end[] __attribute__((weak));
 int initramfs_extract(void) {
     uint8_t *start = _binary_build_initramfs_cpio_start;
     uint8_t *end   = _binary_build_initramfs_cpio_end;
-    uint32_t size  = (uint32_t)(end - start);
+    uint32_t size = (uint32_t)((uintptr_t)end - (uintptr_t)start);
 
     if (size == 0 || size < CPIO_HEADER_SIZE) {
         kprintf("[initramfs] No embedded cpio archive found\n");
